// src/api/generated/model/getApiV1CatCareCatsCatIdWeightRecords200ItemMethod.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * parker-api
 * OpenAPI spec version: 0.1.0
 */

/**
 * @nullable
 */
export type GetApiV1CatCareCatsCatIdWeightRecords200ItemMethod = typeof GetApiV1CatCareCatsCatIdWeightRecords200ItemMethod[keyof typeof GetApiV1CatCareCatsCatIdWeightRecords200ItemMethod] | null;


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiV1CatCareCatsCatIdWeightRecords200ItemMethod = {
  catScale: 'catScale',
  holdAndSubtract: 'holdAndSubtract',
  other: 'other',
} as const;